pyproject.toml
[tool.poetry]
name = "portfolio-manager"
version = "0.1.0"
description = ""
authors = ["zolcsika71 <bushdoctor2008@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "==2.32.3"
python-dotenv = "==1.0.1"
yfinance = "==0.2.50"
tabulate = "==0.9.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
